+++ tests/along_track_etl/test_along_track_etl_integration.py
@@ -1,0 +1,5 @@
+from tests.database.fixtures import *
+
+
+def test_ingest_alongtrack(db_with_alongtrack_data):
+    assert True
