+++ tests/oceandb_initializer/test_oceandb_initializer_unit.py
@@ -1,0 +1,14 @@
+from OceanDB.OceanDB_Initializer import OceanDBInit
+
+from tests.database.fixtures import *
+
+
+def test_basic(config):
+    ocean_db_init = OceanDBInit(config=config)
+    assert ocean_db_init.username == "postgres"
+    assert ocean_db_init.password == "postgres"
+    assert ocean_db_init.port == 5433
+    assert (
+        ocean_db_init.connection_string
+        == "host=localhost dbname=ocean port=5433 user=postgres password=postgres"
+    )
